DOOFUSES Example update to DOTS 1.0 WIP

+++ Example1-DoofusesMustEat/DOOFUSES_UECS/Packages/com.sebaslab.svelto.ecs/Serialization/PartialSerializer.cs
@@ -95,7 +95,7 @@ namespace Svelto.ECS.Serialization
             return true;
         }
 
-        public uint size => totalSize;
+        public int size => (int)totalSize;
 
         static readonly FasterList<(uint, uint)> offsets = new FasterList<(uint, uint)>();
         static readonly uint totalSize;
